deleteDatabase() in other types

+++ app/logic/Calendar/SQL/SQLDatabase.ts
@@ -32,11 +32,16 @@ export async function makeTestDatabase(): Promise<Database> {
   return calendarDatabase;
 }
 
-async function deleteDatabase(): Promise<void> {
+export async function deleteDatabase(): Promise<void> {
   let tables = await calendarDatabase.all(sql`SELECT name FROM sqlite_schema WHERE type='table'`) as any[];
   for (let row of tables) {
     let table = row.name;
+    if (table?.startsWith("sqlite_")) {
+      continue;
+    }
     await calendarDatabase.execute(sql`DROP TABLE IF EXISTS ${table};`);
   }
   await calendarDatabase.pragma('user_version = 0');
+  (calendarDatabase as any).close();
+  calendarDatabase = null;
 }
